# api-contracts/unified-openapi.yaml
openapi: 3.1.0
info:
  title: MyTrader Unified API
  version: 1.1.0
  description: |
    ## MyTrader Trading Platform Unified API Specification

    **CONTRACT VERSION**: 1.1.0
    **EFFECTIVE DATE**: 2024-09-24
    **PRIORITY IMPLEMENTATION**: Phase 1 - WebSocket Standardization, Phase 2 - API Versioning

    ### Critical Issues Addressed:
    1. **Port Configuration**: Standardized on port 5002 (backend) for all environments
    2. **WebSocket Event Naming**: Unified event naming without "Receive" prefixes
    3. **API Versioning**: Consistent `/api/v1/` pattern with backward compatibility
    4. **Volume Leaders Endpoint**: New endpoint for top volume assets per class

    ### Access Patterns
    - **Public Access**: Market data, dashboard, real-time updates (no authentication required)
    - **Authenticated Access**: Trading operations, user profiles, personal data (JWT Bearer token required)

    ### Supported Asset Classes
    - **CRYPTO** - 24/7 trading via multiple data providers
    - **STOCK_BIST** - Borsa Istanbul during Turkish market hours
    - **STOCK_NASDAQ** - US stocks during market hours
    - **FOREX** - Currency pairs (future implementation)

    ### Port Configuration (CRITICAL FIX)
    - **Backend API**: Port 5002 (all environments)
    - **WebSocket Hubs**: Port 5002 (unified with API server)
    - **Frontend Web**: Connects to port 5002 (not 8080)
    - **Frontend Mobile**: Connects to port 5002

...

...

  contact:
...
servers:
...
security:
  - bearerAuth: []
  - {}  # Allow anonymous access for public endpoints
paths:
...
  /symbols:
    get:
      tags: [Symbols]
      summary: Get tracked symbols
      description: Get list of symbols being tracked by the user or system defaults
      security: []
      responses:
        '200':
          description: Tracked symbols retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackedSymbolsResponse'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [Symbols]
      summary: Create new symbol
      description: Add a new symbol to the tracking system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSymbolRequest'
      responses:
        '201':
          description: Symbol created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SymbolResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /symbols/{symbolId}:
    get:
      tags: [Symbols]
      summary: Get symbol by ID
      description: Get detailed information for a specific symbol
      security: []
      parameters:
        - $ref: '#/components/parameters/SymbolIdParam'
      responses:
        '200':
          description: Symbol retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnhancedSymbolResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'

    patch:
      tags: [Symbols]
      summary: Update symbol tracking
      description: Update symbol tracking status and preferences
      parameters:
        - $ref: '#/components/parameters/SymbolIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSymbolRequest'
      responses:
        '200':
          description: Symbol updated successfully
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token obtained from login endpoint
  parameters:
    SymbolIdParam:
      name: symbolId
      in: path
      required: true
      description: Unique symbol identifier
      schema:
        type: string
        format: uuid
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
    UnauthorizedError:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ConflictError:
      description: Resource already exists or conflict occurred
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    CreateSymbolRequest:
      type: object
      required: [ticker, assetClass]
      properties:
        ticker:
          type: string
          minLength: 1
          maxLength: 20
          pattern: '^[A-Z0-9.-]+$'
        assetClass:
          type: string
          enum: [CRYPTO, STOCK_BIST, STOCK_NASDAQ, FOREX]
        market:
          type: string
        baseCurrency:
          type: string
          minLength: 2
          maxLength: 10
        quoteCurrency:
          type: string
          minLength: 2
          maxLength: 10
        displayName:
          type: string
          maxLength: 100
    UpdateSymbolRequest:
      type: object
      properties:
        isTracked:
          type: boolean
        displayName:
          type: string
          maxLength: 100
        notes:
          type: string
          maxLength: 500
    Symbol:
      type: object
      required: [id, ticker, assetClass, isActive]
      properties:
        id:
          type: string
          format: uuid
        ticker:
          type: string
        displayName:
          type: string
        assetClass:
          type: string
          enum: [CRYPTO, STOCK_BIST, STOCK_NASDAQ, FOREX]
        market:
          type: string
        baseCurrency:
          type: string
        quoteCurrency:
          type: string
        isActive:
          type: boolean
        isTracked:
          type: boolean
        tradingHours:
          $ref: '#/components/schemas/TradingHours'
        precision:
          type: object
          properties:
            price:
              type: integer
              minimum: 0
              maximum: 8
            quantity:
              type: integer
              minimum: 0
              maximum: 8
        limits:
          type: object
          properties:
            minTradeAmount:
              type: number
              format: decimal
            maxTradeAmount:
              type: number
              format: decimal
            tickSize:
              type: number
              format: decimal
            lotSize:
              type: number
              format: decimal
        metadata:
          type: object
          properties:
            sector:
              type: string
            industry:
              type: string
            description:
              type: string
            website:
              type: string
              format: uri
    TradingHours:
      type: object
      properties:
        timezone:
          type: string
          example: "America/New_York"
        is24h:
          type: boolean
        regular:
          type: object
          properties:
            monday:
              $ref: '#/components/schemas/DaySchedule'
            tuesday:
              $ref: '#/components/schemas/DaySchedule'
            wednesday:
              $ref: '#/components/schemas/DaySchedule'
            thursday:
              $ref: '#/components/schemas/DaySchedule'
            friday:
              $ref: '#/components/schemas/DaySchedule'
            saturday:
              $ref: '#/components/schemas/DaySchedule'
            sunday:
              $ref: '#/components/schemas/DaySchedule'
    DaySchedule:
      type: object
      properties:
        open:
          type: string
          pattern: '^([01]?[0-9]|2[0-3]):[0-5][0-9]$'
          example: "09:30"
        close:
          type: string
          pattern: '^([01]?[0-9]|2[0-3]):[0-5][0-9]$'
          example: "16:00"
        isClosed:
          type: boolean
    TrackedSymbolsResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                symbols:
                  type: object
                  additionalProperties:
                    $ref: '#/components/schemas/SymbolSummary'
                interval:
                  type: string
                  example: "1m"
                count:
                  type: integer
    SymbolSummary:
      type: object
      required: [symbol, displayName]
      properties:
        symbol:
          type: string
        displayName:
          type: string
        precision:
          type: integer
        strategy_type:
          type: string
    SymbolResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Symbol'
    EnhancedSymbolResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Symbol'
...
    BaseResponse:
      type: object
      required: [success, timestamp]
      properties:
        success:
          type: boolean
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          format: uuid
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            success:
              type: boolean
              enum: [false]
            error:
              type: object
              required: [code, message]
              properties:
                code:
                  type: string
                message:
                  type: string
                details:
                  type: string
                field:
                  type: string
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            success:
              type: boolean
              enum: [false]
            errors:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
...